+++ forms/src/components/index.ts
@@ -1,0 +1,2 @@
+export * from './defaultValidationErrors/defaultValidationErrors.component';
+export * from './defaultValidationErrorsContainer/defaultValidationErrorsContainer.component';