+++ forms/src/misc/tokens.ts
@@ -1,6 +1,8 @@
 import {InjectionToken} from '@angular/core';
 import {StringDictionary} from '@jscrpt/common';
 
+import {ValidationErrorRendererFactoryOptions} from '../services/validationErrorRenderer/validationErrorRenderer.interface';
+
 /**
  * Injection token containing validation error messages
  */
@@ -10,3 +12,8 @@ export const VALIDATION_ERROR_MESSAGES: InjectionToken<StringDictionary> = new I
  * Injection token for injecting array of ignored error names when automatically processed
  */
 export const IGNORED_VALIDATION_ERRORS: InjectionToken<string[]> = new InjectionToken<string[]>('IGNORED_VALIDATION_ERRORS', {providedIn: 'root', factory: () => ['actual']});
+
+/**
+ * Injection token used for injecting global options for ValidationErrorRendererFactory
+ */
+export const VALIDATION_ERROR_RENDERER_FACTORY_OPTIONS: InjectionToken<ValidationErrorRendererFactoryOptions> = new InjectionToken<ValidationErrorRendererFactoryOptions>('VALIDATION_ERROR_RENDERER_FACTORY_OPTIONS');