+++ forms/src/modules/hasError.module.ts
@@ -1,22 +1,30 @@
 import {NgModule} from '@angular/core';
+import {CommonModule} from '@angular/common';
 
 import {HasErrorDirective} from '../directives/hasError/hasError.directive';
 import {GroupHasErrorDirective} from '../directives/groupHasError/groupHasError.directive';
+import {DefaultValidationErrorsComponent, DefaultValidationErrorsContainerComponent} from '../components';
 
 /**
  * Module for input validation directives
  */
 @NgModule(
 {
+    imports:
+    [
+        CommonModule,
+    ],
     declarations: 
     [
         HasErrorDirective,
-        GroupHasErrorDirective
+        GroupHasErrorDirective,
+        DefaultValidationErrorsComponent,
+        DefaultValidationErrorsContainerComponent,
     ],
     exports: 
     [
         HasErrorDirective,
-        GroupHasErrorDirective
+        GroupHasErrorDirective,
     ]
 })
 export class HasErrorModule